requirements.txt
asyncio==3.4.3
buttons==0.1.9
chat-exporter==1.3.0.6.5
discord.py==1.7.3
DiscordUtils==1.2.8
praw==7.4.0
discord-py-slash-command==3.0.
discord-components==2.1.0
dnspython==1.16.0
humanfriendly==9.2
emojis==0.6.0
python-dotenv==0.17.1
motor==2.3.1
Paginator==0.5.1
python-dateutil==2.8.1
python-string-utils==1.0.0
psutil==5.8.0
